Election2005 data

A data set for the German 'Bundestag' election of 2005. It includes information about the elections in 2005 and in 2002 separately for each of the 299 constituencies and also demographic and other information about the constituencies themselves.

data(Election2005)

Format

A data frame with 299 observations on 70 variables. The variables of the data set are:

1-4: general information about the constituencies (ID and name of district, Bundesland)

5-40: demographic and economic information

41 - 48: general information about the elections in 2009 and 2005

WBerechE: number of eligible voters

WE: votes cast

UngZE: invalid second votes

GulZE: valid second votes

49- 70: results of the five biggest parties in 2009 and 2005 (smaller parties are summarized in the variable Rest). Everything with a 'V' or 'v' at the end is from the election in 2005. 'ze' and 'zv' at the end refer to second votes.

Details

This dataset has been taken from the package mbgraphic of Katrin Grimm.

Source

Original source was http://www.bundeswahlleiter.de.

Examples

dim(Election2005)

#> [1] 299  70

str(Election2005)

#> 'data.frame':	299 obs. of  70 variables:
#>  $ PmapNr                       : int  2 1 11 4 7 6 10 13 5 19 ...
#>  $ WahlkreisNr                  : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ Name                         : Factor w/ 299 levels "Aachen","Aalen - Heidenheim",..: 89 201 266 231 145 223 221 254 217 132 ...
#>  $ Land                         : Factor w/ 16 levels "Baden-Wuerttemberg",..: 15 15 15 15 15 15 15 15 15 15 ...
#>  $ Gemeinden...                 : int  137 201 180 164 3 93 49 95 49 126 ...
#>  $ Flaeche.km2.                 : num  2128 2742 2001 2161 143 ...
#>  $ Bevoelk.t.                   : num  286 235 230 252 255 ...
#>  $ maennlich.t.                 : num  141 115 114 124 124 ...
#>  $ Deutsche.t.                  : num  272 226 220 243 232 ...
#>  $ BDichte.je.km2.              : int  134 86 115 116 1785 174 449 223 146 219 ...
#>  $ Bunter15...                  : num  16.7 16.8 17.5 17.3 13.1 16 15.7 16.6 15 15.7 ...
#>  $ B1518...                     : num  3.4 3.3 3.5 3.3 2.5 3.2 3 3.2 3 3.1 ...
#>  $ B1825...                     : num  8 7.3 6.9 7 9.2 6.8 6.7 7.1 6.8 6.5 ...
#>  $ B2535...                     : num  12.9 12.5 12.1 12.3 16.2 12.9 12.2 12.4 11.6 11.6 ...
#>  $ B3560...                     : num  34.4 33.9 35.4 36.2 35.1 35.5 37.5 37.7 35.8 36.7 ...
#>  $ B60mehr...                   : num  24.8 26.2 24.7 23.9 23.9 25.7 24.8 22.9 27.9 26.4 ...
#>  $ LebGeb.je.1000.              : num  8.4 9.2 9.1 8.9 8.6 7.9 8.5 9.1 7.6 8.5 ...
#>  $ Gestorb.je.1000.             : num  11.2 10.5 11.6 10.4 10.1 10.8 9.8 9.6 11.9 10.7 ...
#>  $ Zuzuege.je.1000.             : num  53.7 40.9 33.2 41.8 61.7 46 43.7 46.8 45.6 48.3 ...
#>  $ Fortzuege.je.1000.           : num  47.5 35.5 30.4 38.6 61.2 42.7 37.9 40.5 38.2 41.7 ...
#>  $ Zunahme.je.1000.             : num  3.4 4 0.2 1.7 -1 0.4 4.4 5.7 3.1 4.4 ...
#>  $ LandBetr.je.1000.            : num  9.7 16.8 12.3 10.8 0.4 5.9 4.6 6.1 6.8 4.1 ...
#>  $ LandFl.ha.                   : num  53.6 50.5 47.7 52.7 31 58.7 27.6 46.3 68.8 56.7 ...
#>  $ BergbauBetriebe..je.1000.    : num  0.4 0.4 0.4 0.4 0.4 0.5 0.6 0.7 0.4 0.6 ...
#>  $ BergBauBeschaeftigte.je.1000.: num  44.7 21.7 66.9 28.1 58.1 37.4 54.6 62.8 30.8 61.7 ...
#>  $ NeuWohnungen.je.1000.        : num  4.3 5.5 3.7 3.5 1.7 2.8 4.2 7.4 3.5 5.4 ...
#>  $ Wohnungen.je.1000.           : num  467 516 448 435 548 ...
#>  $ KFZ.je.1000.                 : num  664 728 545 706 526 ...
#>  $ Schulab.je.1000.             : num  11.6 11.3 9.2 9.9 9.6 10.3 10.2 9.8 9.8 9.3 ...
#>  $ ohneHaupt...                 : num  11.5 8.3 8.6 10.9 10.6 10.5 7.8 10.3 12.8 10.5 ...
#>  $ mitHaupt...                  : num  38 43.4 7.6 36.6 26.5 36.7 32.6 33.7 36.7 34.3 ...
#>  $ mitReal...                   : num  27.7 32.6 47.2 31.9 28.7 27.8 32.4 37.4 33.3 30.8 ...
#>  $ mitHoch...                   : num  21.3 14.4 32.7 18.9 31.7 23.1 25.6 15.8 16.2 22.9 ...
#>  $ Arbeitslos03                 : num  11.7 12.8 12.5 8.8 14.4 12 9.8 8.8 12.7 8.6 ...
#>  $ Arbeitslos04                 : num  12.5 13 13.3 10.5 15.4 12.8 9.9 9.1 12.6 8.8 ...
#>  $ SozialVers.je.1000.          : num  268 274 237 229 427 ...
#>  $ LandForstFisch...            : num  2.1 2.3 2.6 3.2 0.2 2.2 3.2 1.5 2.1 1.9 ...
#>  $ Produz...                    : num  24.6 18.5 31.1 25.5 19.4 25.1 32.2 32.2 22.6 35.1 ...
#>  $ Handel...                    : num  27.5 35.1 23.4 25.3 22.2 29 31.5 31.3 32.5 26.9 ...
#>  $ uebDienst...                 : num  45.8 44 42.9 46.1 58.3 43.7 33.1 34.9 42.8 36 ...
#>  $ WBerechE                     : int  220967 184580 177200 197355 194234 174075 227185 233522 177834 232711 ...
#>  $ WBerechV                     : int  217678 181614 175990 195179 193933 172971 223977 228709 175045 228765 ...
#>  $ WE                           : int  171827 143778 139014 158841 151040 136982 185609 188632 140965 190357 ...
#>  $ WV                           : int  171856 144507 140495 159411 152938 139038 187495 187998 141399 190571 ...
#>  $ UngZE                        : int  2596 2058 2108 2062 2048 2100 2287 2107 1842 2458 ...
#>  $ UngZV                        : int  1615 1606 1713 1698 1657 1475 1795 1663 1394 1830 ...
#>  $ GulZE                        : int  169231 141720 136906 156779 148992 134882 183322 186525 139123 187899 ...
#>  $ GulZV                        : int  170241 142901 138782 157713 151281 137563 185700 186335 140005 188741 ...
#>  $ SPDze                        : int  66099 51307 49715 59001 61777 53258 68570 68214 53947 67206 ...
#>  $ SPDzv                        : int  74934 60605 57200 67218 66983 61286 77166 76736 61583 75472 ...
#>  $ CDUCSUze                     : int  61535 59034 53340 59743 42945 48910 67534 68553 52564 69894 ...
#>  $ CDUCSUzv                     : int  62000 57134 53495 58861 42270 49259 65925 68320 53548 71192 ...
#>  $ Grze                         : int  13597 8854 8979 12613 18995 10345 16610 16110 9353 17328 ...
#>  $ GRzv                         : int  15544 9328 10664 14189 22767 11682 18998 17402 9978 18449 ...
#>  $ FDPze                        : int  16397 14270 14701 15507 13190 12945 18796 21167 14442 21046 ...
#>  $ FDPzv                        : int  12688 11721 12250 12665 12530 10549 15419 16193 10963 15412 ...
#>  $ Linkze                       : int  8120 5475 6288 6522 8899 6009 8043 8166 5794 8203 ...
#>  $ Linkzv                       : int  2198 1587 1770 1839 3255 1680 2204 2363 1336 2169 ...
#>  $ SPD                          : num  0.391 0.362 0.363 0.376 0.415 ...
#>  $ CDUCSU                       : num  0.364 0.417 0.39 0.381 0.288 ...
#>  $ Gruene                       : num  0.0803 0.0625 0.0656 0.0805 0.1275 ...
#>  $ FDP                          : num  0.0969 0.1007 0.1074 0.0989 0.0885 ...
#>  $ Linke                        : num  0.048 0.0386 0.0459 0.0416 0.0597 ...
#>  $ Rest                         : num  0.0206 0.0196 0.0284 0.0216 0.0214 ...
#>  $ SPDv                         : num  0.44 0.424 0.412 0.426 0.443 ...
#>  $ CDUCSUv                      : num  0.364 0.4 0.385 0.373 0.279 ...
#>  $ Gruenev                      : num  0.0913 0.0653 0.0768 0.09 0.1505 ...
#>  $ FDPv                         : num  0.0745 0.082 0.0883 0.0803 0.0828 ...
#>  $ Linkev                       : num  0.0129 0.0111 0.0128 0.0117 0.0215 ...
#>  $ Restv                        : num  0.0169 0.0177 0.0245 0.0186 0.023 ...